Remove repeat component
