+++ src/containers/App.jsx
@@ -2,12 +2,12 @@ import "./styles.scss";
 import React from "react";
 import Header from "./Header";
 import { Route, Switch } from "react-router-dom";
-import ProductListing from "./ProductListing";
 import ProductDetails from "./ProductDetails";
 import LoginForm from "./LoginForm";
 import Closet from "./Closet";
 import Hero from "./Hero";
 import Footer from "./Footer";
+import ProductComponent from "./ProductComponent";
 
 const App = () => {
   return (
@@ -15,7 +15,7 @@ const App = () => {
       <Header />
       <Switch>
         <Route path="/" exact component={Hero} />
-        <Route path="/list" exact component={ProductListing} />
+        <Route path="/list" exact component={ProductComponent} />
         <Route path="/login" exact component={LoginForm} />
             <Route path="/login" exact component={LoginForm} />
         <Route path="/:username/inventory" exact component={Closet} />
