--- src/containers/ProductListing.jsx
@@ -1,30 +1,0 @@
-import axios from "axios";
-import React, { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { setProducts } from "../redux/actions/productAction";
-import ProductComponent from "./ProductComponent";
-
-const ProductListing = () => {
-  const dispatch = useDispatch();
-
-  const fetchProducts = async () => {
-    const response = await axios
-      .get(`http://localhost:3001/products`)
-      .catch((error) => {
-        console.log("Error: ", error);
-      });
-    dispatch(setProducts(response.data));
-  };
-
-  useEffect(() => {
-    fetchProducts();
-  }, []);
-
-  return (
-    <div className="list-view">
-      <ProductComponent />
-    </div>
-  );
-};
-
-export default ProductListing;
